% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rich-html-document.R
\name{rich_html_document}
\alias{rich_html_document}
\title{リッチなrmarkdown::html_document}
\usage{
rich_html_document(
  toc = TRUE,
  toc_float = TRUE,
  number_sections = TRUE,
  code_folding = "show",
  code_download = TRUE,
  ...
)
}
\arguments{
\item{toc}{\code{TRUE} to include a table of contents in the output}

\item{toc_float}{\code{TRUE} to float the table of contents to the left of the
main document content. Rather than \code{TRUE} you may also pass a list of
options that control the behavior of the floating table of contents. See the
\emph{Floating Table of Contents} section below for details.}

\item{number_sections}{\code{TRUE} to number section headings}

\item{code_folding}{Enable document readers to toggle the display of R code
chunks. Specify \code{"none"} to display all code chunks (assuming
they were knit with \code{echo = TRUE}). Specify \code{"hide"} to hide all R
code chunks by default (users can show hidden code chunks either
individually or document-wide). Specify \code{"show"} to show all R code
chunks by default.}

\item{code_download}{Embed the Rmd source code within the document and provide
a link that can be used by readers to download the code.}

\item{...}{\code{rmarkdown::html_document}に渡す引数。}
}
\description{
\code{toc}, \code{toc_float}, \code{numbber_sections}, \code{code_folding}, \code{code_download}を
有効にした。
}
