% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styled-word-document.R
\name{spec_reference_docx}
\alias{spec_reference_docx}
\title{オリジナルなWordのテンプレファイルを返す関数}
\usage{
spec_reference_docx()
}
\description{
パッケージのルートディレクトリから見た
\code{inst/docx/reference.docx}
の絶対パスを返す
}
\details{
注意: \code{system.file}は必ず関数の中で使うこと。
CRANなどがバイナリにコンパイルすると、
テンプレートファイルへのパスがコンパイル環境下のものに
固定されてしまう。
}
