% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styled-word-document.R
\name{styled_word_document}
\alias{styled_word_document}
\title{rmarkdown::word_documentにオリジナルテンプレートを適用した版}
\usage{
styled_word_document(reference_docx = spec_reference_docx(), ...)
}
\arguments{
\item{reference_docx}{既定値ならtokyor85downパッケージが用意した
オリジナルのテンプレートファイルを利用する。
それ以外なら指定した値ををそのまま利用する。}

\item{...}{\code{rmarkdown::word_document}に渡す引数。}
}
\description{
rmarkdown::word_documentにオリジナルテンプレートを適用した版
}
