% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styled-html-document.R
\name{styled_html_document}
\alias{styled_html_document}
\title{rmarkdown::html_documentの機能全部有効版}
\usage{
styled_html_document(...)
}
\arguments{
\item{...}{\code{rmarkdown::html_document}に渡す引数。}
}
\description{
\code{toc}, \code{toc_float}, \code{numbber_sections}, \code{code_folding}, \code{code_download}を
有効にした。
}
